Implement direction number recurrence
Follows the math outlined in the paper by Bratley and Fox. There appears
to be an error with the primitive polynomial used to generate the table:
in the paper, it says `x^3 + x + 1`, but the polynomial that matches the
data shown is `x^3 + x^2 + 1`.

There's also a minor issue with the definition of `a_j` by Joe and Kuo,
in that the digits seem to be reversed. They define `a_j` as:

> a_j := (a_{1, j}, a_{2, j}, ..., a_{s_j - 1, j})_2

But the following examples, and the coefficients used by Bratley and
Fox, indicate that the digits are reversed.